feat: drawPepero.swift 파일 생성
기존 main.swift에 작성된 코드에서 새롭게 drawPepero.swift 파일을 생성하여 코드를 분리함
