+++ CodeStarterCamp_Week1/main.swift
@@ -1,31 +1,3 @@
 
-func drawHolder(length: Int) {
-    for _ in 1...length {
-        print(" | | ")
-    }
-}
-
-func choiceBodyAndTopping(body: String, topping: String) -> String {
-    return topping + body + topping
-}
-
-func drawBody(body: String, length: Int) {
-    for _ in 1...length {
-        print(body)
-    }
-}
-
-func drawPepero(holderLength: Int, body: String, topping: String = " ", peperoLength: Int) {
-    print("""
-    <정보>
-    길이: \(peperoLength)
-    몸통: \(body)
-    토핑: \(topping)
-    막대길이: \(holderLength)
-    
-    """)
-    drawBody(body: choiceBodyAndTopping(body: body, topping: topping), length: peperoLength)
-    drawHolder(length: holderLength)
-}
-
+// Week1 STEP3 - 주문에 따른 빼빼로 그리기
 drawPepero(holderLength: 4, body: "|0|", topping: "#", peperoLength: 10)
